Improve Structure
